+++ edgy/core/db/models/mixins/date_parser.py
@@ -1,0 +1,16 @@
+import typing
+from enum import Enum
+
+from orjson import OPT_OMIT_MICROSECONDS, OPT_SERIALIZE_NUMPY, dumps
+
+
+class DateParser:
+    def _resolve_value(self, value: typing.Any) -> typing.Any:
+        if isinstance(value, dict):
+            return dumps(
+                value,
+                option=OPT_SERIALIZE_NUMPY | OPT_OMIT_MICROSECONDS,
+            ).decode("utf-8")
+        elif isinstance(value, Enum):
+            return value.name
+        return value
